move common/state to app/state
