+++ pkg/window/toolwindow/projectexplorer/projectexplorer.go
@@ -2,6 +2,7 @@
 package projectexplorer
 
 import (
+	"github.com/gucio321/HellSpawner/pkg/app/state"
 	"log"
 	"os"
 	"path/filepath"
@@ -19,7 +20,6 @@ import (
 	"github.com/gucio321/HellSpawner/pkg/common/hsfiletypes"
 	"github.com/gucio321/HellSpawner/pkg/common/hsproject"
 	"github.com/gucio321/HellSpawner/pkg/common/hsutil"
-	"github.com/gucio321/HellSpawner/pkg/common/state"
 	"github.com/gucio321/HellSpawner/pkg/widgets"
 	"github.com/gucio321/HellSpawner/pkg/window/toolwindow"
 )
