+++ pkg/window/toolwindow/toolwindow.go
@@ -2,7 +2,7 @@ package toolwindow
 
 import (
 	"github.com/AllenDang/giu"
-	"github.com/gucio321/HellSpawner/pkg/common/state"
+	"github.com/gucio321/HellSpawner/pkg/app/state"
 	"github.com/gucio321/HellSpawner/pkg/window"
 )
